+++ MoreShipUpgrades/Managers/PatchManager.cs
@@ -112,6 +112,7 @@ namespace MoreShipUpgrades.Managers
             harmony.PatchAll(typeof(WalkiePatcher));
             harmony.PatchAll(typeof(SprayPaintItemPatcher));
             harmony.PatchAll(typeof(JetpackItemPatcher));
+            harmony.PatchAll(typeof(LungPropPatcher));
             Plugin.mls.LogInfo("Items have been patched");
         }
         static void PatchVitalComponents()