+++ MoreShipUpgrades/Plugin.cs
@@ -23,7 +23,7 @@ namespace MoreShipUpgrades
     [BepInPlugin(Metadata.GUID, Metadata.NAME, Metadata.VERSION)]
     [BepInDependency(LethalLib.Plugin.ModGUID, "0.13.0")]
     [BepInDependency("com.sigurd.csync")]
-    [BepInDependency(LethalCompanyInputUtils.PluginInfo.PLUGIN_GUID)]
+    [BepInDependency(LethalCompanyInputUtils.MyPluginInfo.PLUGIN_GUID)]
     [BepInDependency("WhiteSpike.InteractiveTerminalAPI")]
     [BepInDependency(CustomItemBehaviourLibrary.Misc.Metadata.GUID, DependencyFlags.SoftDependency)]
     [BepInDependency(LethalLevelLoader.Plugin.ModGUID, DependencyFlags.SoftDependency)]