+++ MoreShipUpgrades/UpgradeComponents/Items/MonsterSample.cs
@@ -1,5 +1,6 @@
 ﻿using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
+using MoreShipUpgrades.UpgradeComponents.TierUpgrades.Ship;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -27,8 +28,15 @@ namespace MoreShipUpgrades.UpgradeComponents.Items
                 usedMapSeed = StartOfRound.Instance.randomMapSeed;
                 random = new System.Random(usedMapSeed + 105);
             }
-            GetComponent<ScrapValueSyncer>().SetScrapValue(random.Next(minValue: itemProperties.minValue, maxValue: itemProperties.maxValue));
-            particles = GetComponentInChildren<ParticleSystem>();
+			int value = random.Next(minValue: itemProperties.minValue, maxValue: itemProperties.maxValue);
+			if (UpgradeBus.Instance.PluginConfiguration.AffectSamples)
+			{
+				Plugin.mls.LogDebug($"Midas Touch affecting samples is enabled, increasing the original scrap value ({value})...");
+				value = MidasTouch.IncreaseScrapValueInteger(value);
+				Plugin.mls.LogDebug($"Set the sample scrap value to {value}...");
+			}
+			GetComponent<ScrapValueSyncer>().SetScrapValue(value);
+			particles = GetComponentInChildren<ParticleSystem>();
         }
         public override void EquipItem()
         {