+++ ChangeLog.md
@@ -2,7 +2,7 @@
 <summary> 3.12.13.2 - 2026-??-?? </summary>
 
 - Added code related to LethalLevelLoader's ExtendedLevels as DunGen is no longer included in ``AssemblyCSharp.dll`` but rather its own assembly (``DunGen.dll``).
-
+- Made Midas Touch influence both samples dropped by creatures through Hunter upgrade and the apparatices that spawn in the interior's special rooms (both toggleable through configuration)
 </details>
 
 <details>